refactor: separate elemtent edit modal submission logic from modal and form
